// src/LoDb.Web/src/app/core/api/generated/models/summoner-facets.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * The facet options the data decides, over the whole list.
 */
export interface SummonerFacets {

  /**
   * The unlock levels, ascending.
   */
  levels: Array<number>;

  /**
   * The filterable modes some spell allows, in the facet's display order.
   */
  modes: Array<string>;
}